edited navbar, cards, small style stuff
