+++ my-app/client/src/App.js
@@ -3,7 +3,8 @@ import './App.css';
 import Navbar from './components/Navbar';
 import { BrowserRouter as Router, Routes, Route}
     from 'react-router-dom';
-import Home from './Pages/home';
+import Home from './Pages/';
+import About from './Pages/about';
 import Login from './Pages/Login';
 import Donate from './Pages/donate';
 import Favedogs from './Pages/favedogs';
@@ -15,7 +16,8 @@ return (
     <Router>
     <Navbar />
     <Routes>
-        <Route exact path='/home' element={<Home />} />
+        <Route exact path='/' element={<Home />} />
+        <Route path='/about' element={<About/>} />
         <Route path='/favedogs' element={<Favedogs/>} />
         <Route path='/signup' element={<Signup/>} />
         <Route path='/login' element={<Login/>} />
